backend/requirements.txt
django==5.0rc1
djangorestframework==3.14.0
django-cors-headers==4.3.1
py-encry==0.2.0
